Migrate get_history from utils to __init__
