+++ browser_history/cli.py
@@ -3,7 +3,7 @@ command line interface of browser-history."""
 
 import sys
 import argparse
-from browser_history import generic, browsers, utils
+from browser_history import __init__, generic, browsers
 
 # get list of all implemented browser by finding subclasses of generic.Browser
 AVAILABLE_BROWSERS = ', '.join(b.__name__ for b in generic.Browser.__subclasses__())
@@ -39,7 +39,7 @@ def main():
     args = parser.parse_args()
 
     if args.browser == 'all':
-        outputs = utils.get_history()
+        outputs = __init__.get_history()
 
     else:
         try:
